Add content based duplicate detection
The MD5 based duplicate detection is not viable solution as it pollutes
the ledger file with meta data that is meaningless for everything but
the icsv2 importer.
This change adds content based duplicate detection to the script. When
starting up we read the current content of the ledger file and remember
that. When a new entry is about to be added we simply search for this
new entry in a byte-by-byte fashion.